requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.48.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==72.2.0
six==1.16.0
soupsieve==2.6
urllib3==2.2.2
Werkzeug==3.0.4
